requirements.txt
pandas>=2.2.0
openpyxl>=3.1.0
pyarrow>=15.0.0
numpy>=1.25.0
requests>=2.32.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
scikit-learn>=1.4.0
matplotlib>=3.7.0
seaborn>=0.12.2
